python-runtime/pyproject.toml
[project]
name = "python-runtime"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "grpcio>=1.67.0",
    "grpcio-tools>=1.67.0",
]
